requirements.txt
fastapi>=0.111.0,<0.120.0
uvicorn>=0.29.0,<0.35.0
groq>=0.9.0,<0.20.0
supabase>=2.4.6,<3.0.0
pydantic>=2.7.0,<3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
streamlit>=1.35.0,<2.0.0